requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-swagger==2.2.0
django-rq==2.3.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mercadopago==1.1.1
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
